Delete README.md
To play the game, the files may be downloaded on your personal computer in a directory or folder called "Game." Then inside of the terminal or console, run the following command inside quotations: "g++ *.cpp". Then type "./a.out" and press enter. The game should run and the main menu will be displayed with the game rules, the player settings, and the start/exit of the game.